Update pyo3 dependency to latest version (#255)
Updates the PyO3 dependency in the `python_ext` crate to 0.27.1.
Thankfully all the deprecations triggered in the update are trivial to
fix.

This also fixes installing `llguidance` on the free-threaded build of
Python 3.14. Free-threaded Python doesn't support abi3, so the existing
wheels on PyPI don't work.

It'd also be really nice to get a release out with this change because
it would fix building this project on 3.14t and unblock the possibility
of releasing wheels that support Python 3.14t.

This would ultimately partially unblock vLLM adding support for the
free-threaded build: https://github.com/vllm-project/vllm/issues/28762.