Mark immutable pyclasses as frozen (#260)
See [the PyO3 docs](https://pyo3.rs/v0.27.1/class.html#frozen-classes-opting-out-of-interior-mutability) for more about the `frozen` option for the pyclass macro.

In short, it tells PyO3 that the Rust data wrapped in the class cannot be mutated via a `&mut` reference. It also makes it a compiler error to create mutable references. It's possible in the future `frozen` will become the default.

The main benefit is performance. Frozen pyclasses do not need to do any atomic reference counting when rust data gets borrowed. It also makes these types a little easier to reason about in a multithreaded context. Since all these types can be trivially marked as `frozen`, the default atomic reference counting scheme in PyO3 is unnecessary.

It's possible the other two pyclasses in the library could be marked as `frozen` as well with a little more work by leveraging some form of interior mutability.